test(cli): keep pagination coverage within lint limits
